added docstrings to views files, improved venue creation on admin page

+++ website/shop/views.py
@@ -10,6 +10,11 @@ shop_blueprint = Blueprint("shop", __name__, template_folder="/templates")
 
 @shop_blueprint.route("/browse", methods=["POST", "GET"])
 def browse():
+    """Loads browse page to display all purchasable tickets, manages sort_button and search_bar
+
+    Renders:
+        browse.html: on load
+    """
     global browseConcerts
 
     match request.form.get("sort_button"):
@@ -64,6 +69,11 @@ def browse():
 @shop_blueprint.route("/cart", methods=["GET", "POST"])
 @login_required
 def cart():
+    """Loads the cart page to show the current users unpurchased tickets.
+
+    Renders:
+        cart.html: on load
+    """
     cart = (
         Ticket.query.filter_by(ownerId=current_user.id).filter_by(purchased=False).all()
     )
@@ -74,10 +84,16 @@ def cart():
 @shop_blueprint.route("/add_to_cart", methods=["POST"])
 @login_required
 def add_to_cart():
+    """Creates ticket for concert and adds it to cart.
+
+    Redirects:
+        shop.cart: on load
+    """
     concert = Concert.query.filter_by(id=request.form.get("purchase_button")).first()
 
     if concert:
-        concert.create_ticket(ownerId=current_user.id)
+        if concert.availableTickets > 0:
+            concert.create_ticket(ownerId=current_user.id)
 
     return redirect(url_for("shop.cart"))
 
@@ -85,7 +101,11 @@ def add_to_cart():
 @shop_blueprint.route("/remove_from_cart", methods=["POST"])
 @login_required
 def remove_from_cart():
+    """Deletes ticket from cart.
 
+    Redirects:
+        shop.cart: on load
+    """
     ticket = Ticket.query.filter_by(
         id=request.form.get("remove_from_cart_button")
     ).first()
@@ -99,12 +119,17 @@ def remove_from_cart():
 @shop_blueprint.route("/empty_cart", methods=["GET", "POST"])
 @login_required
 def empty_cart():
+    """Deletes all tickets from cart
+
+    Redirects:
+        shop.cart: on load
+    """
     cart = (
         Ticket.query.filter_by(ownerId=current_user.id).filter_by(purchased=False).all()
     )
-
-    for ticket in cart:
-        ticket.delete()
+    if cart:
+        for ticket in cart:
+            ticket.delete()
 
     return redirect(url_for("shop.cart"))
 
@@ -112,13 +137,19 @@ def empty_cart():
 @shop_blueprint.route("/buy_additional_ticket", methods=["POST"])
 @login_required
 def buy_additional_ticket():
+    """Adds new ticket to cart for the same concert as
+
+    Redirects:
+        shop.cart: on load
+    """
     ticket = Ticket.query.filter_by(
         id=request.form.get("buy_additional_ticket_button")
     ).first()
 
     if ticket:
-        concert = ticket.get_concert()
-        concert.create_ticket(ownerId=current_user.id)
+        if ticket.get_concert().availableTickets > 0:
+            concert = ticket.get_concert()
+            concert.create_ticket(ownerId=current_user.id)
 
     return redirect(url_for("shop.cart"))
 
@@ -126,6 +157,14 @@ def buy_additional_ticket():
 @shop_blueprint.route("/purchase", methods=["GET", "POST"])
 @login_required
 def purchase():
+    """Loads purchase page, allows user to input payment details, if successful purchases all tickets in cart.
+
+    Renders:
+        reciept.html: on successful purchase
+        purchase.html: on unsuccessful purchase
+    Errors:
+        403: Forbidden
+    """
     cart = (
         Ticket.query.filter_by(ownerId=current_user.id).filter_by(purchased=False).all()
     )
@@ -155,7 +194,7 @@ def purchase():
             cart=cart,
             totalPrice=totalPrice,
         )
-        
+
     else:
         return abort(403, "Forbidden")
 
@@ -163,11 +202,16 @@ def purchase():
 @shop_blueprint.route("/view_ticket", methods=["GET", "POST"])
 @login_required
 def view_ticket():
+    """Creates QR code and renders ticket page from view_ticket_button
+
+    Renders:
+        ticket.html: if ticket exists
+    Errors:
+        403: Forbidden
+    """
     ticket = Ticket.query.filter_by(id=request.form.get("view_ticket_button")).first()
 
     if ticket:
-        confirmationCode = ticket.confirmationCode
-
         qr = qrcode.QRCode(
             version=1,
             error_correction=qrcode.constants.ERROR_CORRECT_L,
@@ -175,7 +219,7 @@ def view_ticket():
             border=2,
         )
 
-        qr.add_data(confirmationCode)
+        qr.add_data(ticket.confirmationCode)
         qr.make(fit=True)
         img = qr.make_image(fill_color="black", back_color="white")
         img.save("website/static/qr.png")
@@ -183,8 +227,8 @@ def view_ticket():
         return render_template(
             "ticket.html",
             ticket=ticket,
-            confirmationCode=confirmationCode,
+            confirmationCode=ticket.confirmationCode,
         )
-        
+
     else:
         return abort(403, "Forbidden")
